Fix: intergrate grpc
[x] Intergrate grpc server to runner service
[x] Implement grpc in tester service
[x] Fix calling stat from tester to docker

+++ tester/app/models/Record.py
@@ -7,6 +7,6 @@ class Record(BaseModel):
     total: int
     passed: int
     status_list : list[Status]
-    time_response : list[int]
+    time_response : list[float]
     memory_used : list[int]
